# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Next Step Sports Platform API
  version: 1.0.0-draft
  description: |
    Initial agent-build contract for the Next Step Sports Platform.
    This contract is a starting skeleton and must remain aligned with the
    business/product specification and accepted ADRs.
servers:
  - url: https://api.staging.nextstep.example/v1
    description: Staging
  - url: http://localhost:3001/v1
    description: Local
security:
  - oidc: []
...
paths:
...
  /me:
    get:
      tags: [Identity]
      operationId: getCurrentUser
      summary: Return the authenticated user and authorised household context
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        default:
          $ref: "#/components/responses/Problem"
  /households/{householdId}/athletes:
    parameters:
      - $ref: "#/components/parameters/HouseholdId"
    get:
      tags: [Athletes]
      operationId: listAthletes
      summary: List active and optionally archived athletes in the household
      parameters:
        - name: includeArchived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Athlete list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Athlete"
        default:
          $ref: "#/components/responses/Problem"
    post:
      tags: [Athletes]
      operationId: createAthlete
      summary: Create a parent-controlled private athlete profile
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAthleteRequest"
      responses:
        "201":
          description: Athlete created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Athlete"
        default:
          $ref: "#/components/responses/Problem"
...
  /athletes/{athleteId}:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    get:
      tags: [Athletes]
      operationId: getAthlete
      responses:
        "200":
          description: Athlete profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Athlete"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      tags: [Athletes]
      operationId: updateAthlete
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/UpdateAthleteRequest"
      responses:
        "200":
          description: Athlete updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Athlete"
        default:
          $ref: "#/components/responses/Problem"
  /athletes/{athleteId}/baseline:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    post:
      tags: [Athletes]
      operationId: submitBaseline
      summary: Submit baseline inputs and assign a recommended starting campaign
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BaselineRequest"
      responses:
        "200":
          description: Baseline result and campaign recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaselineResult"
        default:
          $ref: "#/components/responses/Problem"
  /athletes/{athleteId}/dashboard:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    get:
      tags: [Progress]
      operationId: getAthleteDashboard
      summary: Return the athlete's single primary next action and concise progress context
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AthleteDashboard"
        default:
          $ref: "#/components/responses/Problem"
  /athletes/{athleteId}/skill-tree:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    get:
      tags: [Progress]
      operationId: getSkillTree
      parameters:
        - name: domainKey
          in: query
          schema:
            type: string
        - name: stageKey
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Athlete-specific curriculum tree and progress state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillTree"
        default:
          $ref: "#/components/responses/Problem"
  /athletes/{athleteId}/skills/{nodeId}:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
      - $ref: "#/components/parameters/NodeId"
    get:
      tags: [Progress]
      operationId: getAthleteSkillDetail
      summary: Return contextual skill content, progress and remaining requirements
      responses:
        "200":
          description: Athlete-specific skill detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AthleteSkillDetail"
        default:
          $ref: "#/components/responses/Problem"
...
  /athletes/{athleteId}/practice-plans:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    post:
      tags: [Practice]
      operationId: generatePracticePlan
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePracticePlanRequest"
      responses:
        "201":
          description: Persisted practice plan snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticePlan"
        default:
          $ref: "#/components/responses/Problem"
  /practice-plans/{planId}:
    parameters:
      - $ref: "#/components/parameters/PlanId"
    get:
      tags: [Practice]
      operationId: getPracticePlan
      responses:
        "200":
          description: Practice plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticePlan"
        default:
          $ref: "#/components/responses/Problem"
  /practice-plans/{planId}/sessions:
    parameters:
      - $ref: "#/components/parameters/PlanId"
    post:
      tags: [Practice]
      operationId: startPracticeSession
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                clientSessionId:
                  type: string
                startedOffline:
                  type: boolean
                  default: false
      responses:
        "201":
          description: Practice session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSession"
        default:
          $ref: "#/components/responses/Problem"
  /practice-sessions/{sessionId}/attempts:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [Practice]
      operationId: recordPracticeAttempt
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeAttemptRequest"
      responses:
        "201":
          description: Attempt recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeAttempt"
        default:
          $ref: "#/components/responses/Problem"
  /practice-sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [Practice]
      operationId: getPracticeSession
      summary: Retrieve resumable practice-session state
      responses:
        "200":
          description: Current session and recorded attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSessionDetail"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      tags: [Practice]
      operationId: updatePracticeSession
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/UpdatePracticeSessionRequest"
      responses:
        "200":
          description: Session updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSession"
        default:
          $ref: "#/components/responses/Problem"
  /practice-sessions/{sessionId}/complete:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [Practice]
      operationId: completePracticeSession
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletePracticeSessionRequest"
      responses:
        "200":
          description: Completion result and progress changes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeCompletionResult"
        default:
          $ref: "#/components/responses/Problem"
  /evidence/upload-intents:
    post:
      tags: [Evidence]
      operationId: createEvidenceUploadIntent
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUploadIntentRequest"
      responses:
        "201":
          description: Short-lived private upload intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadIntent"
        default:
          $ref: "#/components/responses/Problem"
  /evidence/upload-intents/{mediaAssetId}/complete:
    parameters:
      - $ref: "#/components/parameters/MediaAssetId"
    post:
      tags: [Evidence]
      operationId: completeEvidenceUpload
      summary: Finalise a direct private upload and queue media validation
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteEvidenceUploadRequest"
      responses:
        "202":
          description: Upload accepted for private validation and processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAsset"
        default:
          $ref: "#/components/responses/Problem"
  /evidence-submissions:
    post:
      tags: [Evidence]
      operationId: submitEvidence
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitEvidenceRequest"
      responses:
        "201":
          description: Evidence submitted and assessment requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceSubmission"
        default:
          $ref: "#/components/responses/Problem"
...
  /coach/assessment-queue:
    get:
      tags: [Assessments]
      operationId: getCoachAssessmentQueue
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Assigned assessments ordered by due time
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AssessmentSummary"
                  nextCursor:
                    type: [string, "null"]
        default:
          $ref: "#/components/responses/Problem"
...
  /coach/assessments/{assessmentId}/decision:
    parameters:
      - $ref: "#/components/parameters/AssessmentId"
    post:
      tags: [Assessments]
      operationId: submitAssessmentDecision
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentDecisionRequest"
      responses:
        "200":
          description: Assessment completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResult"
        default:
          $ref: "#/components/responses/Problem"
...
  /athletes/{athleteId}/passport:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    get:
      tags: [Passport]
      operationId: getAthletePassport
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: eventType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Private athlete passport and timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AthletePassport"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/.well-known/openid-configuration
  parameters:
    HouseholdId:
      name: householdId
      in: path
      required: true
      schema: { type: string, format: uuid }
    AthleteId:
      name: athleteId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PlanId:
      name: planId
      in: path
      required: true
      schema: { type: string, format: uuid }
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { type: string, format: uuid }
    NodeId:
      name: nodeId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    MediaAssetId:
      name: mediaAssetId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    AssessmentId:
      name: assessmentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
  responses:
    Problem:
      description: Problem details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    CurrentUser:
      type: object
      required: [id, roles, households]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: [string, "null"] }
        roles:
          type: array
          items: { type: string }
        households:
          type: array
          items:
            type: object
            required: [id, role]
            properties:
              id: { type: string, format: uuid }
              name: { type: [string, "null"] }
              role: { type: string }
    Athlete:
      type: object
      required: [id, householdId, displayName, ageBand, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        householdId: { type: string, format: uuid }
        displayName: { type: string, minLength: 1, maxLength: 80 }
        ageBand: { type: string, examples: ["U10"] }
        birthMonth: { type: [integer, "null"], minimum: 1, maximum: 12 }
        birthYear: { type: [integer, "null"], minimum: 2000, maximum: 2030 }
        laterality: { type: [string, "null"] }
        status: { type: string, enum: [ACTIVE, ARCHIVED, DELETION_PENDING] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateAthleteRequest:
      type: object
      required: [displayName, ageBand, consents]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 80 }
        ageBand: { type: string }
        birthMonth: { type: integer, minimum: 1, maximum: 12 }
        birthYear: { type: integer, minimum: 2000, maximum: 2030 }
        laterality: { type: string }
        consents:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ConsentGrant"
    UpdateAthleteRequest:
      type: object
      properties:
        displayName: { type: string, minLength: 1, maxLength: 80 }
        ageBand: { type: string }
        laterality: { type: [string, "null"] }
        status: { type: string, enum: [ACTIVE, ARCHIVED] }
...
    BaselineRequest:
      type: object
      required: [sportKey, experienceLevel, environment, observations]
      properties:
        sportKey: { type: string, const: basketball }
        experienceLevel: { type: string }
        environment:
          $ref: "#/components/schemas/PracticeEnvironment"
        observations:
          type: array
          items:
            type: object
            required: [key, value]
            properties:
              key: { type: string }
              value: {}
    BaselineResult:
      type: object
      required: [athleteId, curriculumVersion, recommendedStageKey, campaign, explanation]
      properties:
        athleteId: { type: string, format: uuid }
        curriculumVersion: { type: string }
        recommendedStageKey: { type: string }
        campaign:
          $ref: "#/components/schemas/CampaignSummary"
        explanation:
          type: array
          items: { type: string }
        mayChooseMoreFoundationalStart: { type: boolean, const: true }
    AthleteDashboard:
      type: object
      required: [athlete, primaryAction, campaign, weeklySummary]
      properties:
        athlete:
          $ref: "#/components/schemas/Athlete"
        primaryAction:
          $ref: "#/components/schemas/NextAction"
        campaign:
          $ref: "#/components/schemas/CampaignSummary"
        weeklySummary:
          type: object
          required: [meaningfulPractices, progressing]
          properties:
            meaningfulPractices: { type: integer, minimum: 0 }
            progressing: { type: boolean }
            revisitDueCount: { type: integer, minimum: 0 }
        pendingAssessment:
          oneOf:
            - $ref: "#/components/schemas/AssessmentSummary"
            - type: "null"
...
    SkillTree:
      type: object
      required: [curriculumVersion, campaign, domains, nodes]
      properties:
        curriculumVersion: { type: string }
        campaign:
          $ref: "#/components/schemas/CampaignSummary"
        domains:
          type: array
          items:
            type: object
            required: [key, name, sortOrder, completedNodeCount, totalNodeCount, progress]
            properties:
              key: { type: string }
              name: { type: string }
              sortOrder: { type: integer }
              completedNodeCount: { type: integer, minimum: 0 }
              totalNodeCount: { type: integer, minimum: 0 }
              progress: { type: number, minimum: 0, maximum: 1 }
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/AthleteSkillNode"
        branchChoices:
          type: array
          items:
            $ref: "#/components/schemas/BranchChoice"
...
    AthleteSkillDetail:
      allOf:
        - $ref: "#/components/schemas/AthleteSkillNode"
        - type: object
          required: [objective, whyItMatters, childCues, safety, primaryAction]
          properties:
            objective: { type: string }
            whyItMatters: { type: string }
            childCues:
              type: array
              maxItems: 3
              items: { type: string }
            commonErrors:
              type: array
              items: { type: string }
            safety:
              type: array
              items: { type: string }
            primaryAction:
              $ref: "#/components/schemas/NextAction"
...
    GeneratePracticePlanRequest:
      type: object
      required: [durationPreset, environment, recentLoad]
      properties:
        durationPreset: { type: string, enum: [QUICK, STANDARD, EXTENDED] }
        environment:
          $ref: "#/components/schemas/PracticeEnvironment"
        selectedBranchNodeId: { type: [string, "null"], format: uuid }
        recentLoad:
          type: object
          required: [fatigue, pain]
          properties:
            fatigue: { type: string, enum: [LOW, NORMAL, HIGH] }
            pain: { type: boolean }
            note: { type: [string, "null"], maxLength: 1000 }
    PracticePlan:
      type: object
      required: [id, athleteId, title, purpose, status, targetDurationMinutes, generationReasons, steps, createdAt]
      properties:
        id: { type: string, format: uuid }
        athleteId: { type: string, format: uuid }
        title: { type: string }
        purpose: { type: string }
        status: { type: string, enum: [GENERATED, STARTED, COMPLETED, EXPIRED, CANCELLED] }
        targetDurationMinutes: { type: integer, minimum: 5, maximum: 60 }
        generationReasons:
          type: array
          items: { type: string }
        steps:
          type: array
          items:
            $ref: "#/components/schemas/PracticePlanStep"
        createdAt: { type: string, format: date-time }
        expiresAt: { type: [string, "null"], format: date-time }
...
    PracticeSession:
      type: object
      required: [id, practicePlanId, athleteId, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        practicePlanId: { type: string, format: uuid }
        athleteId: { type: string, format: uuid }
        status: { type: string, enum: [NOT_STARTED, IN_PROGRESS, COMPLETED, ABANDONED, SAFETY_STOPPED] }
        startedAt: { type: [string, "null"], format: date-time }
        completedAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        version: { type: integer, minimum: 1, default: 1 }
    PracticeSessionDetail:
      allOf:
        - $ref: "#/components/schemas/PracticeSession"
        - type: object
          required: [attempts]
          properties:
            attempts:
              type: array
              items:
                $ref: "#/components/schemas/PracticeAttempt"
    UpdatePracticeSessionRequest:
      type: object
      properties:
        status: { type: string, enum: [IN_PROGRESS, ABANDONED, SAFETY_STOPPED] }
    PracticeAttemptRequest:
      type: object
      required: [planStepId, attemptNumber, resultType, skipped]
      properties:
        planStepId: { type: string, format: uuid }
        attemptNumber: { type: integer, minimum: 1 }
        resultType: { type: string, enum: [BOOLEAN, COUNT, DURATION, SUCCESS_RATIO, OBSERVATION, NONE] }
        result: {}
        difficultyRating: { type: [integer, "null"], minimum: 1, maximum: 5 }
        cueUnderstood: { type: [boolean, "null"] }
        skipped: { type: boolean }
        skipReason: { type: [string, "null"], maxLength: 500 }
        completedAt: { type: string, format: date-time }
    PracticeAttempt:
      allOf:
        - $ref: "#/components/schemas/PracticeAttemptRequest"
        - type: object
          required: [id, sessionId]
          properties:
            id: { type: string, format: uuid }
            sessionId: { type: string, format: uuid }
    CompletePracticeSessionRequest:
      type: object
      required: [enjoymentRating, difficultyRating, safetyFlag]
      properties:
        enjoymentRating: { type: integer, minimum: 1, maximum: 5 }
        difficultyRating: { type: integer, minimum: 1, maximum: 5 }
        parentObservation: { type: [string, "null"], maxLength: 1000 }
        safetyFlag: { type: boolean }
        safetyNote: { type: [string, "null"], maxLength: 1000 }
        completedAt: { type: string, format: date-time }
    PracticeCompletionResult:
      type: object
      required: [session, progressChanges, nextAction]
      properties:
        session:
          $ref: "#/components/schemas/PracticeSession"
        progressChanges:
          type: array
          items:
            type: object
            required: [nodeId, priorState, newState, reason]
            properties:
              nodeId: { type: string, format: uuid }
              priorState: { $ref: "#/components/schemas/ProgressState" }
              newState: { $ref: "#/components/schemas/ProgressState" }
              reason: { type: string }
        nextAction:
          $ref: "#/components/schemas/NextAction"
    CreateUploadIntentRequest:
      type: object
      required: [athleteId, nodeId, consentRecordId, filename, mimeType, sizeBytes]
      properties:
        athleteId: { type: string, format: uuid }
        nodeId: { type: string, format: uuid }
        consentRecordId: { type: string, format: uuid }
        filename: { type: string, maxLength: 255 }
        mimeType: { type: string }
        sizeBytes: { type: integer, minimum: 1 }
        durationMs: { type: [integer, "null"], minimum: 1, maximum: 90000 }
        checksumSha256: { type: [string, "null"] }
    UploadIntent:
      type: object
      required: [mediaAssetId, uploadUrl, expiresAt, requiredHeaders]
      properties:
        mediaAssetId: { type: string, format: uuid }
        uploadUrl: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
        requiredHeaders:
          type: object
          additionalProperties: { type: string }
        multipart:
          type: boolean
          default: false
    CompleteEvidenceUploadRequest:
      type: object
      required: [checksumSha256]
      properties:
        checksumSha256: { type: string, minLength: 64, maxLength: 64 }
        uploadId: { type: [string, "null"] }
        parts:
          type: array
          items:
            type: object
            required: [partNumber, etag]
            properties:
              partNumber: { type: integer, minimum: 1 }
              etag: { type: string }
...